packages/dataset_cli/pyproject.toml
[project]
name = "davis-cli"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "polars>=1.32.3",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "typer>=0.16.1",
    "weasyprint>=65.1",
    "jinja2>=3.1.6",
    "charset-normalizer>=3.4.2",
    "dvc[gdrive]>=3.62.0",
]

[project.scripts]
davis = "dataset_cli.main:app"

[tool.uv]
package = true

[dependency-groups]
dev = ["types-pyyaml>=6.0.12.20250822"]

[tool.setuptools.package-data]
dataset_cli = ["locales/*/LC_MESSAGES/*.mo"]

[tool.uv.scripts]
post-install = "uv run pybabel compile"
